+++ src/opts.rs
@@ -13,10 +13,30 @@ pub struct Opts {
     /// $XDG_CACHE_HOME/break-time/ (or ~/.cache/break-time/ if $XDG_CACHE_HOME is not set).
     #[structopt(long, name = "CACHE_DIR_PATH", parse(from_os_str))]
     pub cache_dir: Option<PathBuf>,
+
+    #[structopt(subcommand)]
+    pub cmd: Option<Command>,
 }
 
 impl Opts {
     pub fn parse_from_args() -> Self {
         Self::from_args()
     }
+}
+
+#[derive(Debug, StructOpt)]
+pub enum Command {
+    GoogleCalendar(GoogleCalendar),
+}
+
+#[derive(Debug, StructOpt)]
+pub enum GoogleCalendar {
+    ListEvents,
+    IgnoreEvent(IgnoreEvent),
+}
+
+#[derive(Debug, StructOpt)]
+pub struct IgnoreEvent {
+    /// Event ID.  You can get this with `break-time google-calendar list-events`.
+    pub event_id: String,
 }